refactor(configure): extract block availability into a hook
Pulls the palette's "which factories can join the canvas" memo into
useAvailableFactoryIds() so the command palette can apply the same rule.
