+++ frontend/src/features/fable-builder/components/layout/BlockPalette.tsx
@@ -25,6 +25,10 @@ import type {
 } from '@/api/types/fable.types'
 import { useFableBuilderStore } from '@/features/fable-builder/stores/fableBuilderStore'
 import {
+  isFactoryAvailable,
+  useAvailableFactoryIds,
+} from '@/features/fable-builder/hooks/useAvailableFactoryIds'
+import {
   BLOCK_KIND_METADATA,
   BLOCK_KIND_ORDER,
   factoryIdToKey,
@@ -65,35 +69,9 @@ export function BlockPalette({ catalogue }: BlockPaletteProps) {
     (state) => state.setDraggedFactory,
   )
   const fable = useFableBuilderStore((state) => state.fable)
-  const validationState = useFableBuilderStore((state) => state.validationState)
   const isValidating = useFableBuilderStore((state) => state.isValidating)
 
-  const blockCount = Object.keys(fable.blocks).length
-
-  const availableFactoryIds = useMemo(() => {
-    if (blockCount === 0) {
-      if (!validationState) {
-        // Validation not available yet — signal sources-only mode.
-        // TODO: Change when backend validation works properly
-        return 'sources-only' as const
-      }
-      return new Set(
-        validationState.possibleSources.map((id) => factoryIdToKey(id)),
-      )
-    }
-
-    if (!validationState) return null
-
-    const allExpansions = new Set<string>()
-    for (const blockState of Object.values(validationState.blockStates)) {
-      for (const expansion of blockState.possibleExpansions) {
-        allExpansions.add(factoryIdToKey(expansion))
-      }
-    }
-    // No expansions (e.g. a block has errors) → keep every block available
-    // instead of greying the palette. Mirrors AddNodeButton's fallback.
-    return allExpansions.size > 0 ? allExpansions : null
-  }, [validationState, blockCount])
+  const availableFactoryIds = useAvailableFactoryIds()
 
   const groupedFactories = useMemo(() => {
     const groups = new Map<
@@ -126,11 +104,11 @@ export function BlockPalette({ catalogue }: BlockPaletteProps) {
           factory: factoryId,
         }
         const key = factoryIdToKey(pluginBlockFactoryId)
-        const isAvailable =
-          availableFactoryIds === null ||
-          (availableFactoryIds === 'sources-only'
-            ? factory.kind === 'source'
-            : availableFactoryIds.has(key))
+        const isAvailable = isFactoryAvailable(
+          availableFactoryIds,
+          factory,
+          key,
+        )
         group.push({ id: pluginBlockFactoryId, factory, isAvailable })
       }
     }
